### 2019-02-08 #281 @DillonHammill
Submission: CytoRSuite - Flow Cytometry Analysis Toolkit
Submitting Author Name: Dillon Hammill
Submitting Author Github Handle: @DillonHammill
Repository:   https://github.com/DillonHammill/CytoRSuite
Version submitted:   0.9.9
Version accepted: TBD

---

```
Package: CytoRSuite
Title: Compensation, Gating & Visualisation Toolkit for Analysis of Flow Cytometry Data
Version: 0.9.9
  person("Dillon", "Hammill", role = c("aut", "cre"), email = "Dillon.Hammill@anu.edu.au"),
  person("Greg", "Finak", role = "ctb", email = "gfinak@fredhutch.org"),
  person("Mike", "Jiang", role = "ctb", email = "wjiang2@fhcrc.org"))
Description: An intuitive and interactive approach to flow cytometry analysis in R.
  CytoRSuite contains tools for every stage of the analysis pipeline including
  user-guided automatic compensation, realtime editing of spillover matrices,
  manual gate drawing, gate editing, visualisation of all existing flow cytometry
  objects and export of population level statistics for further analysis. Furthermore,
  through linking to openCyto, CytoRSuite provides the tools necessary to easily combine
  manual and automated gating approaches in R.
  R (>= 2.10),
  openCyto(>= 1.21.1),
  flowWorkspace (>= 3.28.2),
  flowCore (>= 1.49.4)
  ncdfFlow,
  BiocGenerics (>= 0.29),
  shiny,
  shinythemes,
  rhandsontable,
  MASS,
  graphics,
  grDevices,
  tools,
  magrittr
  CytoRSuiteData,
  RProtoBufLib,
  cytolib,
  vdiffr,
  shinytest,
  covr,
  robustbase,
  pkgdown
  Bioconductor/BiocGenerics@a2de4dd,
  RGLab/RProtoBufLib@trunk,
  RGLab/flowCore@trunk,
  RGLab/ncdfFlow@trunk,
  RGLab/cytolib@trunk,
  RGLab/flowWorkspace@trunk,
  RGLab/openCyto@trunk,
  DillonHammill/CytoRSuiteData@master
```

CytoRSuite uses the existing flow cytometry infrastructure (flowCore and flowWorkspace packages) to import flow cytometry data. CytoRSuite (which is built on using the openCyto package) provides the toolkit necessary to interactively manipulate, process and visualise this flow cytometry data.

CytoRSuite is primarily targeted to scientists who traditionally utilise software with a GUI to manually gate their flow cytometry data. CytoRSuite aims to form a bridge between this GUI software and R by providing a similar user experience with minimal coding.  Furthermore, CytoRSuite aims to provide a means by which manual gating users can be exposed to and eased into automated gating approaches.

Although other R packages which support automated gating approaches exist, to the best of my knowledge there is no R package which supports manual gating. CytoRSuite not only brings the ability to manually draw gates, but also the potential to combine automated and manual gating approaches through the openCyto gatingTemplate. Furthermore, CytoRSuite expands the existing compensation tools by providing an interactive shiny application to fine tune spillover matrix values in realtime, Additionally, CytoRSuite contains an intuitive set of plotting functions which can provide powerful visualisations with minimal coding.  CytoRSuite integrates all the tools necessary for the analysis of flow cytometry data in a single cohesive package.
