### 2019-02-25 #281 @geanders
---

@DillonHammill : My apologies for the delays in getting this out for review! It looks like a very interesting package, and I'm very excited to be serving as an editor. Please let me know throughout the process if you have any questions. Also, I am much less familiar with Bioconductor than I am with the CRAN ecosystem, so I may have some misunderstandings along the way. Please let me know if there's something I've misunderstood, especially related to Bioconductor, for this package.

In the next few days, I will be assigning reviewers for the package. In the meantime, here are a few small issues that preliminary checks brought up that you could address in the package:

**Some tips for reviewers**

The README for the package's GitHub repository has some very nice instructions for required BioConductor and GitHub dependencies. As a note to reviewers, make sure you read through these, and you can use the code examples in [the README](https://github.com/DillonHammill/CytoRSuite) to install those dependencies, especially if you clone the `CytoRSuite` package code and build locally rather than using `devtools::install_github` to get it and its dependencies.

Also, it seems that on MacOS, users might need to install XQuartz to allow the `vdiffr` package to load correctly. I got the same error message described [here](https://github.com/thomasp85/gganimate/issues/95) when trying to install the package. This was resolved once I installed XQuartz on my computer and rebooted.

- [ ] @DillonHammill, could you clarify if the reviewers need to install each of the Bioconductor dependencies from Bioconductor, from GitHub, or both? Instructions for both are given on the README for each package, but perhaps just the latest from Bioconductor (or just the latest from GitHub) would suffice?
- [ ] Also, will the reviewers need to install `CytoRSuiteData` before they can successfully install `CytoRSuite`, since the data for `CytoRSuite`'s vignettes are in the data package? If so, could you make this note for the reviewer, and consider changing the order of installation instructions in the package's GitHub README.
- [ ] Does the latest version of `BiocGenerics` also need to be installed? If so, it would be worthwhile to add that with the other Bioconductor instructions in the README, as well as where to find the right version (it looks like it needs commit "a2de4dd" from Bioconductor's GitHub).
- [ ] The XQuartz dependency may be worth checking up on a bit more, as well as adding a note in the README.

**`spelling::spell_check_package()` output**

- Could you add the line `Language: en-GB` to the package's "DESCRIPTION" file so that British spellings aren't flagged? (Otherwise, the default for spell checking is to assume US spellings)
- The spellcheck flagged a few spelling errors that could be legitimate:

```
indicies                    gate_draw-flowSet-method.Rd:15
                                 gate_draw-GatingSet-method.Rd:23
                                 gate_edit.Rd:14
recommmend          spillover_compute-flowSet-method.Rd:16
suppply                    CytoRSuite.Rmd:254
```

- [ ] Fix any valid spelling typos

**`goodpractice::gp()` output**

The following issues came up for `goodpractice` checks:

```

✖ write unit tests for all functions, and all package code in
    general. 70% of code lines are covered by test cases.
```

- [ ] It would be useful to increase the coverage of the test cases.

```

  ✖ write short and simple functions. These functions have high
    cyclomatic complexity:gate_web_draw (113).
```

- [ ] If reasonable, simplify the `gate_web_draw` function, possibly breaking it into a few simpler functions.

```

  ✖ fix this R CMD check ERROR: Package suggested but not
    available: ‘pkgdown’ The suggested packages are required for a complete
    check. Checking can be attempted without them by setting the environment
    variable _R_CHECK_FORCE_SUGGESTS_ to a false value. See section ‘The
    DESCRIPTION file’ in the ‘Writing R Extensions’ manual.
```

I am not completely sure why this error is coming up, because `pkgdown` is available on CRAN. It would be great to resolve this if we can.

```

  ✖ not use "Depends" in DESCRIPTION, as it can cause name clashes,
    and poor interaction with other packages. Use "Imports" instead.
```
- [ ] Change from "Depends" to "Imports" in the "DESCRIPTION" file. For more on why this is considered good practice, see the Chapters in the "R Packages" book on [Package metadata](http://r-pkgs.had.co.nz/description.html) and [Namespaces](http://r-pkgs.had.co.nz/namespace.html).

```
  ✖ not import packages as a whole, as this can cause name clashes
    between the imported packages. Instead, import only the specific
    functions you need.
```

- [ ] Change to import specific functions from a package rather than the whole package. Fixing the previous issue (using "Depends" rather than "Imports" when possible) may fix this issue. You could run `goodpractice::gp()` on the package after change to "Depends" to confirm that.

```
  ✖ avoid 'T' and 'F', as they are just variables which are set to
    the logicals 'TRUE' and 'FALSE' by default, but are not reserved words
    and hence can be overwritten by the user.  Hence, one should always use
    'TRUE' and 'FALSE' for the logicals.
```

- [ ] Change `T` / `F` to `TRUE` / `FALSE` throughout.

---
