### 2019-02-27 #281 @DillonHammill
Thanks for the feedback @geanders.

I have addressed the following today:
1. Updated the installation instructions in the README to install flowCore, flowWorkspace and openCyto development versions from BioConductor. These versions are synced with the GitHub versions.
2. Changed the installation order to first install CytoRSuiteData and then CytoRSuite.
3. BiocGenerics is required by the package dependencies, however I think the most recent version on GitHub requires the development version of R. I will check to see if this is still the case.
4. XQuartz is required for the functioning of CytoRSuite, it is responsible for opening interactive plotting windows to facilitate gate drawing on Mac. Windows and Linux machines will use the equivalent windows() and X11() functions from grDevices.
5. Added en-GB to language in DESCRIPTION.
6. Fix some spelling errors.

Comments on `goodpractice()`:
1. I will try to include more unit tests for the package, I have been trying to stick to a 1 minute run for unit tests. If the test running time is not an issue I will aim to add more. There is also some testing for the shiny application `spillover_edit` which is not included in the code coverage (~5%).
2. `gate_web_draw` is quite a complex function, I will see what I can do to simplify it.
3. Not sure about the `pkgdown` error, I don't seem to get this on my windows machine.
4. In terms of package dependencies, flowCore, flowWorkspace and openCyto contain a lot of the tools required at various stages throughout the analysis pipeline. These packages are maintained by the same group to ensure that there are no conflicts. For this reason I have added these to Depends field of the DESCRIPTION file.
5. I have also had to import the entire `grDevices` library as I require the platform-specific functions `windows()`, X11() and `quartz()`. If I explicitly import `windows()` or `X11()` this will cause problems on a mac where these functions are not included in the grDevices libraries.
6. I haven't used any T/F throughout the package. I have checked everything and I have no idea why this is showing in `goodpractice()`.

Questions:
1. I was wondering if I am still able to submit the package to BioConductor as well?
2. I am working on a publication, when this is complete can it be used for submission to JOSS?

Thanks for your help.
Dillon
