### 2019-03-25 #281 @DillonHammill
@geanders, sorry for not getting back to you sooner.

CytoRSuite is definitely designed to be used in conjunction with flowCore, flowWorkspace and OpenCyto (as in the above description). In particular, CytoRSuite builds on OpenCyto which depends on both flowCore and flowWorkspace. I written CytoRSuite in such way as to prevent clashes with these otrher packages.

I will start working on a paper for JOSS submission in the meantime.

Any news on the reviewer front?
