### 2019-03-28 #281 @jacobpwagner
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

Estimated hours spent reviewing: ~5

---

CytoRSuite adds very helpful interactive interfaces over the core infrastructure of flowCore, flowWorkspace, and openCyto. The point-and-click interface for gate construction (`gate_draw()` methods) and the ability to easily see the effects of changes to spillover matrix entries (`spillover_edit()`) greatly enhance the approachability and usability of the RGLab packages. This is important for users performing a good amount of manual gating or more accustomed to analysis using a GUI.

Additionally, the plots generated by `cyto_plot()` are both useful and aesthetically pleasing.
It's clear that you took some care to choose a vibrant set of color schemes that would yield good
contrast and highlight density nicely. Other multiple-plotting methods (i.e. `cyto_plot_profile()`)
are similarly helpful.

For the most part, everything ran smoothly. However, I ran in to a few minor issues:

---

Your vignettes cover all of the major functionality of the package and are very helpful.
Some of the example code does run in to some problems when run by the user:

* In `CytoRSuite-Compensation.Rmd`, it looks like you used the wrong name
for the GatingTemplate filename argument to `gate_draw()`in the example code
(`gtfile` instead of `gatingTemplate`). The vignette still builds correctly
because you use the pre-saved png, but users walking through your code won't
correctly save out "Compensation gatingTemplate.csv". There are a few such lines.

* Similarly, in `CytoRSuite-Manual-Automate-Gating.Rmd`, the example code has a
has a few lines with calls to `app_pop`, which I assume is meant to be `add_pop`.

* In `CytoRSuite-Gating-Functions.Rmd`, the example code coerces a `flowSet` to
a `flowFrame` (line 70), then calls `gate_polygon_draw` on that `flowFrame`. This
results in an error:
```
Error in if (is.na(title)) { : argument is of length zero
In addition: Warning messages:
1: In rep(x[[arg]], length.out = plots) :
  'x' is NULL so the result will be NULL
2: In rep(x[[arg]], length.out = n) :

```
You can avoid the error by providing a dummy "title" argument, but ultimately
this is coming down to line 2472 in `cyto_plot_internal.R` where in the absence
of a title, you assign it like so:
```
if (missing(title)) {
  title <- fr@description$GUID
}
```
The problem is that the `flowCore` method for coercing a `flowSet` to a `flowFrame`
does not assign a value to `description$GUID`, so you end up with `NULL`. There may
be other related issues beyond the problem for the vignette, as a quick grep shows
that there are 15 other lines directly accessing the `GUID` element from the
`description` slot of an object in this `@description$GUID`manner. You should
just make sure that the object in each case will actually have that. A better
approach may be using `flowCore::identifier()` which will do those checks for you
and at least give you "anonymous" if it doesn't find anything.

* `CytoRSuite.Rmd` and `CytoRSuite-Gating-Functions.Rmd` both mention sealing
polygon gates in `gate_draw()` by right-clicking and selecting "stop". The
"stop" button doesn't seem to be present on Linux (I just closed the window)
but that's really minor and I'm guessing users can figure it out.

---

The code is generally very readable and  if anything over-commented so it was easy
to follow and generally seems sound. A few minor issues:

* I was working with a fresh R 3.5.3 installation on Linux and `X11.options()$type` was
defaulting to "Xlib" instead of "cairo" (which should be the default when available), which meant that your `X11()` calls resulting from any case where `popup == TRUE` in `cyto_plot()` were yielding windows for which `dev.capabilities()$semiTransparency` was `FALSE`. That had a bunch of downstream effects due to `alpha` arguments, including warnings ("semi-transparency is not supported on this device") and disabling some features like the plots in the "Plots" tab of the Spillover Matrix Editor. While I quickly figured out the cause and fixed it by `X11.options(type="cairo")`, the message and problems are unclear enough that other users may get confused.
While you could leave this up to users to troubleshoot, you could also always add a check for `dev.capabilities()$semiTransparency == TRUE` or `grepl("cairo",X11.options()$type)` before `X11()` calls. I think everything may filter down to the single `X11()` line in `.cyto_plot_window()` so guarding that with a more helpful warning may be one option.

* If a `gate_draw()` window is closed without the user actually building a gate, it
produces an error:
```
Error in locator(type = "o", lwd = 2, pch = 16, col = "red") :
  graphics device closed during call to locator or identify
```
In terms of end result, it seems fine. No gate is added, which is what I'd expect.
However, maybe it would be good to catch/suppress the error for a graceful exit.

* In the `cyto_plot()` methods, there are at least two layers of function calls. First,
a `cyto_plot()` layer that mostly determines whether it should call the 1D or 2D internal
plotting method, and then the appropriate 1D or 2D internal method. Each of these layers
provides the full set of default arguments and `cyto_plot()` passes each along in a long call
to `cyto_plot_1d()` or `cyto_plot_2d()` but it looks like essentially the only argument
that is relevant to the first layer is `channels`. For the sake of code maintenance it
may be easier if you can make it work by just using the `channels` argument in the `cyto_plot()`
layer and just passing the rest along to `.cyto_plot_1d()` or `.cyto_plot_2d()` in `...`. Then you would have only one copy of the list of default arguments. But you may have a reason for doing it this way.

* There are a few methods in `helper-functions.R` that include an `else` case with some apparently magic numbers.
I'm thinking of something like line 229 in `cyto_channel_select()`:
```
...
      if (getOption("CytoRSuite_interact") == TRUE) {
        menu(choices = opts, graphics = TRUE)

        # Test channels - 7AAD, AF430, APC Cy7, NIL, PE Cy7, PE
        c(4, 7, 11, 12, 5, 2)[match(x, pData(fs)$name)]
      }
...
```
I'm guessing this is related to some of the test data you use in the vignettes from `CytoRSuiteData`,
but it seems like it may be poor practice to hard-code it in to a more general method, especially if
it should be throwing an error for being called in a non-interactive context. If you need
this default for testing, there may be a better way. For example, you could alter the method to allow
specification of a set of channels. But again, there may be some other reason for this.

* `CytoRSuite::spillover_compute()` has a good amount of overlap with `flowCore::spillover()` (the `flowSet` method). There are some significant peripheral differences, but I will probably run a few test cases to make sure the core computation is coming out the same as users may expect this.

---

Your documentation looks nice and thorough and the links work well for making navigation easy. I verified that all exported methods have documentation with examples and ran a fair amount of the examples to make sure they worked as expected. R CMD check didn't raise any warnings or errors. My only suggestion would be to maybe break up some of the longer "Details" sections in to a few paragraphs to improve readability (`spillover_edit` for example).

---
Other than that, the package looks great. I may continue to run it through some testing and I will update this with any issues I run in to. I look forward to seeing what else you will be adding and please feel free to contact me with issues of flowCore/flowWorkspace/openCyto compatibility.
