### 2019-03-28 #281 @DillonHammill
Thanks for taking the time to review CytoRSuite @jacobpwagner. I will post some comments and make the appropriate changes tomorrow.
