### 2019-03-30 #281 @DillonHammill
@jacobpwagner thanks again for your help!

I have made a number of changes to CytoRSuite over the last couple of days. I have committed these to the `testing` branch until I am convinced that everything is stable. Please install from this branch to see if I have resolved the issues you have raised.

Changes:
- Updated Compensation vignette to remove remaining old argument names (e.g. gtfile)
- Fixed typing error in Manual/Automated gating vignette (add_pop)
- Fixed gating functions to prevent title and legend_text plotting errors by switching over to using `identifier()`.
- Added a check to `X11()` calls for semiTransparency functionality. If this capability is missing from the graphics device `X11.options()$type` is set to "cairo". I have also included a reset to return this to the default on exit. I am unable to test this properly on my windows machine so please let me know if your `X11.options()` are in fact returned to the default once the plotting window is closed.
- All gating functions should now have a graceful exit upon closing the gating window without drawing any gates.
- A number of improvements to `cyto_plot()` to set better axes limits to display all events correctly.
- Made some aesthetic changes to the plots tab of `spillover_edit()` to improve visibility.

New features & Improvements:
- `gate_draw()`, `gate_edit()` and `gate_remove()` are now completely interchangeable. This means that users can simply change "draw" to "edit" or "remove" to modify gates.
- Added ability to automatically offset labels in `cyto_plot()` when multiple gates are drawn close together.
- Added the ability to supply empty alias (e.g. "") to automatically plot all gates constructed in the supplied channels.
- Added new methods for computing spillover spreading matrices (`spillover_spread`) which are still being actively developed and tested. I still need to add examples and documentation for these methods.
- Added a sample size restriction to `gate_draw ` to improve processing speed with large flowSet or GatingSet objects. This has been to set to an excessive 20 flowFrames, which will be fine tuned over time.

Comments:
- For `cyto_plot` I have designed it to have all arguments readily available in the documentation (no searching required) but also for autocomplete. I noticed that (...) arguments were not being displayed in the autocomplete list. Happy to hear better alternatives.
- Writing unit tests for the package has been challenging as there are a lot of interactive functions which require user input. The best approach I could come up with was to create a interactive mode (`options(CytoRSuite_interact = FALSE)`) which is switched off during unit testing. The "magic" numbers are random user selections which are used internally to test these interactive functions. If you know of a better way of doing this please let me know.
- `spillover_compute()` is similar to the flowCore variety but has some differences. There used to be a problem with normalisation used internally in `spillover()` but I think this has now been resolved? Also `spillover_compute()` calculates spillover to unused channels as well which didn't happen previously with `spillover()`. I will also be adding option to use an internal negative population for each compensation control as the reference for the calculation. Happy to review `spillover()` to further investigate differences.

I will start working on a paper for JOSS submission this week.

Please let me know if you encounter any additional issues.
