### 2019-04-09 #281 @jacobpwagner
@DillonHammill, here are my comments after the recent changes
### Testing changes to fix issues raised
* The warnings when a user just closes an X11 window from `gate_draw()` instead of completing a gate are gone. -- FIXED
* The switch to using identifier() for the specified lines seems to have fixed the issue in the vignette resulting from the `flowSet` coercion. -- FIXED

* Vignette typos -- FIXED

* For the X11 issue, you can actually just use `X11(type="cairo")` rather than temporarily changing it using `X11.options()` and changing it back. I'm sorry. I was really unclear before. I just meant more that you could add a check of whether cairo is available, as users can compile R without it. But you could even just add it as a requirement to your docs just as you did for quartz for macOS. The alternative is guarding plotting calls with `alpha` args using `suppressWarnings`.

  The solution of https://github.com/DillonHammill/CytoRSuite/commit/53bf3883b541e90385d0f3f1fe34e4a5973a2b7e actually won't work as the `dev.capabilities$semiTransparency == TRUE` check happens before the `X11()` call, so it will check the semi-transparency of the currently active device (e.g. RStudioGD) before opening the window. I'm sorry I didn't think about that when I mentioned it as a potential solution in my prior comment. Anyway, you can probably leave your code as it was before that commit but instead switch `X11()` to `X11(type="cairo")`. I've tried that and it works fine.

* For the unit testing user input issue, there are a few ways around the interaction problem, but one is to just locally mock the interactive function (e.g. `menu()`) to provide the specified input (the hard-coded test columns of your else clauses). `test_that::with_mock()` accomplishes this but can't be used for functions in base packages (like `utils::menu()`), but the `mockery` package (or `mockr`) makes it easy to work around this. Either way, that pulls the testing code out of the source code for your package. For example, you have a test that looks like this:
```
test_that("cyto_channel_select", {
  print(pData(Comp))
  expect_equal(cyto_channel_select(Comp[[1]]), "PE-Cy7-A"))
  expect_equal(cyto_channel_select(Comp),
               c("7-AAD-A",
                 "Alexa Fluor 430-A",
                 "APC-Cy7-A",
                 "Unstained",
                 "PE-Cy7-A",
                 "PE-A"))
  expect_equal(cyto_channel_select(gsc),
               c("7-AAD-A",
                 "Alexa Fluor 430-A",
                 "APC-Cy7-A",
                 "Unstained",
                 "PE-Cy7-A", "PE-A"))
})
```
  depending on the test code in the else clauses of the conditionals on "CytoRSuite_interact" in the method definitions:
```
setMethod(cyto_channel_select,
          signature = "flowFrame",
          definition = function(x) {

            # Assign x to fr
            fr <- x

            opts <- cyto_fluor_channels(fr)

            # Print sample name and select channel
            message(
              paste(
                "Select a fluorescent channel for the following csample:",
                fr@description$GUID
              )
            )

            if (getOption("CytoRSuite_interact") == TRUE) {
              channel <- opts[menu(choices = opts, graphics = TRUE)]
            } else {
              # Tests use PE Cy7 Control -
              channel <- opts[5]
            }

            return(channel)
          }
)

```
Instead, you could just mock the `menu` call using `mockery::stub()` or `mockery::mock()`. I've just changed over the first check here after getting rid of the "CytoRSuite_interact" conditional with the test-case code for the flowFrame method and it passes just fine.
```
test_that("cyto_channel_select", {
  print(pData(Comp))
  mock_menu <- mock(5)
  with_mock(menu = mock_menu, expect_equal(cyto_channel_select(Comp[[1]]), "PE-Cy7-A"))
  expect_equal(cyto_channel_select(Comp),
               c("7-AAD-A",
                 "Alexa Fluor 430-A",
                 "APC-Cy7-A",
                 "Unstained",
                 "PE-Cy7-A",
                 "PE-A"))
  expect_equal(cyto_channel_select(gsc),
               c("7-AAD-A",
                 "Alexa Fluor 430-A",
                 "APC-Cy7-A",
                 "Unstained",
                 "PE-Cy7-A", "PE-A"))
})
```
where now the method just looks like this (no more hard-coded numbers for testing):
```
setMethod(cyto_channel_select,
          signature = "flowFrame",
          definition = function(x) {

            # Assign x to fr
            fr <- x

            opts <- cyto_fluor_channels(fr)

            # Print sample name and select channel
            message(
              paste(
                "Select a fluorescent channel for the following csample:",
                fr@description$GUID
              )
            )
            channel <- opts[menu(choices = opts, graphics = TRUE)]
            return(channel)
          }
)

```
Side note: That `fr@description$GUID` may be another good candidate for `identifier(fr)`, but I didn't really explore that.

### Other comments
* The logic for the `cyto_plot` argument lists makes sense. I thought there was probably a good reason and I agree that it is good to make the arguments available for autocomplete. It's up to you whether you want to comment the deeper layer of default args to make it clear that they will be overridden by the first layer.

* No worries about the differences between `spillover_compute()` and `flowCore::spillover()`. I just thought I'd note it as a reminder (mostly to myself) to test their agreement in more detail. But the method looks sound.

* I haven't had time to exhaustively check the new features and improvements, but if I encounter any issues I will let you know.

### Summary
Most of the issues I raised were addressed. The X11 issue is still there and is the only thing I pointed out that still really needs to be fixed, but the fix is pretty easy as I discussed. It's up to you whether you want to pull the test code out of your methods as I described, but as a general habit it will make it easier for collaborators to work with improving your methods if they do not have test code incorporated in them.
