### 2019-04-09 #281 @DillonHammill
@jacobpwagner thanks again for your help!

I have pushed the fix for `X11()` which should now work as expected.

I did not know about `mockr` I will look at updating my unit tests in this manner later today. I will also do a complete sweep of the package to replace all instances of `fr@description$GUID` with `identifier()`.
