### 2019-04-16 #281 @lmweber
Hi @DillonHammill , thank you for submitting this package. My review comments are pasted below:

---

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

**Overall comments**: This package provides a set of tools for analyzing flow cytometry data in R, specifically for compensation, interactive gating, and visualization. This fills a clear need in the flow cytometry data analysis community for performing manual gating and related tasks more easily within R, instead of using commercial platforms (currently, tasks such as pre-processing and more automated analysis pipelines are often performed in R, while manual gating and related tasks are usually done using commercial platforms such as Cytobank and FlowJo). The package has extensive documentation, including several vignettes with examples of workflows. The package is also well integrated with existing R tools for analyzing flow cytometry data, including flowCore, flowWorkspace, and openCyto. I think this will be a very useful package for the flow and mass cytometry community, and am happy to have had a chance to look at it in some more detail.

However, I ran into several errors and other issues when trying to run the examples in the vignettes. These are described below, along with some more general suggestions.

---

**README**

Currently, the README includes the same content as the [index page of the main documentation](https://dillonhammill.github.io/CytoRSuite/index.html). By having the same content in two places, there is a risk that they will become out of sync in future. You could consider consolidating this by shortening the README and including some links to the main documentation pages instead.

It would also be useful to update the README to include some information on the 'Statement of need' (see checklist above), e.g., mentioning where CytoRSuite fits into the set of available analysis tools, and the intended audience. This is described nicely in the 'Scope' section in the first comment in this issue (8 Feb), so you could re-use some of that text.

---

**Scope**

One related tool that is not mentioned in the 'Scope' section above is iSEE, available from [Bioconductor](http://bioconductor.org/packages/iSEE). iSEE is intended more generally for interactive visualization of single-cell data stored in SummarizedExperiment objects, but one of their vignettes shows how to use it for gating mass cytometry data, so there is some overlap there. However, CytoRSuite contains much more specialized functionality for flow cytometry data and builds on existing flow cytometry packages such as flowCore / flowWorkspace / openCyto, so I don't think this overlap reduces the usefulness of CytoRSuite. Both tools provide very useful functionality for users who wish to perform their full analysis pipelines within R, instead of switching to commercial products for the manual gating steps.

You could also consider mentioning (either in the README or in the main documentation) something about how this fits in with tools for analyzing higher-dimensional flow and mass cytometry data (e.g., CytoRSuite could be used for initial pre-gating and for visualizations).

---

**Installation and checks**

I installed CytoRSuite on a Mac, following the instructions on the README and main documentation pages. The installation completed without errors.

I also ran `devtools::check()`, however this returned a few errors and notes, which should be addressed.

- I first ran `devtools::check()` on the `testing` branch, since this is the most current version (according to the previous review comments above). However, this gave an ERROR in building one of the vignettes:

```
Quitting from lines 83-84 (CytoRSuite-Visualisation.Rmd)
   Error: processing vignette 'CytoRSuite-Visualisation.Rmd' failed with diagnostics:
   argument is of length zero
```

Due to this error, I switched to the `master` branch to run the rest of the checks. This still gave the following NOTES:

- NOTE

```
N  checking package dependencies (2.2s)
   Packages suggested but not available for checking:
     ‘vdiffr’ ‘shinytest’ ‘covr’ ‘pkgdown’
```

These packages are listed in the `Suggests:` field in the DESCRIPTION, so I thought they would be installed automatically when using `devtools::install_github()` with `dependencies = TRUE`. I'm not sure why this failed. After installing the packages manually, the NOTE was removed. This would probably also be fixed by distributing the CytoRSuite package through Bioconductor.

- NOTE

```
N  checking installed package size ...
       doc   28.2Mb
       help   3.1Mb
```

The largest files are the `.gif` files in `docs/articles/Gating-functions`. I think this is probably fine; although you could consider reducing the size of these files, e.g., by downsampling the resolution, as long as this does not affect the visual quality in the documentation.

- NOTE

```
N  checking R code for possible problems (33.9s)
   Found an obsolete/platform-specific call in the following function:
     ‘.cyto_plot_window’
   Found the platform-specific devices:
     ‘X11’ ‘quartz’
   dev.new() is the preferred way to open a new device, in the unlikely
   event one is needed.
   Found the following assignments to the global environment:
   File ‘CytoRSuite/R/gatingTemplate-modifiers.R’:
     assign(deparse(substitute(x)), gs, envir = globalenv())
     assign(deparse(substitute(x)), gs, envir = globalenv())
   File ‘CytoRSuite/R/helper-functions.R’:
     assign(deparse(substitute(x)), cyto, envir = globalenv())
```

I'm less sure about these; it would be good if they can be addressed, but I don't have any specific advice about how to fix them.

---

I tested the functionality by trying to run through the examples in the vignettes. However, I ran into a number of problems, which will need to be addressed:

- Using the `testing` branch, I had the following ERROR when running functions such as `gate_draw()`:

```
Error in if (getOption("CytoRSuite_cyto_plot_grid")) { :
  argument is of length zero
```

Due to this error, I switched back to the `master` branch. However, this still gave several warnings and errors, including:

- Warnings in vignette 'Compensation of Fluorescent Spillover', section '2.3 Gate Compensation Controls', after creating the gate with function `gate_draw()`:

```
1: In plot.window(...) : "gtfile" is not a graphical parameter
2: In plot.xy(xy, type, ...) : "gtfile" is not a graphical parameter
3: In axis(side = side, at = at, labels = labels, ...) :
  "gtfile" is not a graphical parameter
4: In axis(side = side, at = at, labels = labels, ...) :
  "gtfile" is not a graphical parameter
5: In box(...) : "gtfile" is not a graphical parameter
6: In title(...) : "gtfile" is not a graphical parameter
```

- Error in vignette 'Gating Functions', section '3.2 gate_polygon_draw', when running the function `gate_polygon_draw()`:

```
Error in if (is.na(title)) { : argument is of length zero
In addition: Warning messages:
1: In rep(x[[arg]], length.out = plots) :
  'x' is NULL so the result will be NULL
2: In rep(x[[arg]], length.out = n) :
  'x' is NULL so the result will be NULL
```

Also, an additional minor issue was that at first, I didn't realize that right-clicking was required to complete the polygons and continue to the next step of the gating. I think this is mentioned in some of the code comments, but it would be good to mention it more prominently at the start.

I am happy to try running these vignette examples again once the errors and warnings have been addressed.

---

As mentioned above, the package includes extensive vignettes, which is great for new users. The package also includes help files for the individual functions.

One suggestion I have for the function help files is regarding places where you have several help files covering very similar content for related functions (e.g., `gate_polygon_draw()`, `gate_rectangle_draw()`, `gate_boundary_draw()`). For these functions, you could consolidate the help files into a single page using `@rdname` in Roxygen (e.g., see the instructions in section 'Documenting multiple functions in the same file' in [Hadley Wickham's R Packages book](http://r-pkgs.had.co.nz/man.html)). This would make it easier for you to maintain, and simpler for the reader to get an overview.

I also ran into a problem with the examples in the first vignette (i.e., the examples in the README and on the index page of the main documentation). These examples begin with the code:

```
library(CytoRSuite)
library(CytoRSuiteData)

# Save .fcs files to folder "Compensation Controls" in working directory
files <- list.files(path = "Compensation Controls", full.names = TRUE)
fs <- read.ncdfFlowSet(files = files)
```

If I understand it correctly, the intention here is that the reader saves some `.fcs` files in the directory `Compensation Controls`, and then continues with the example. However, it would be much more useful if some `.fcs` files were provided along with the package. Alternatively, you could update these examples to use one of the datasets from the `CytoRSuiteData` package, similar to the other vignettes.

---

**Code quality**

The code looks well-structured and commented. If you also submit the package to Bioconductor, you will need to check that you meet Bioconductor's guidelines regarding code style.

---

**Additional suggestions**

**Bioconductor**: I would strongly recommend also submitting this package to Bioconductor. The package already builds on other Bioconductor packages, so this should be feasible. Making the package available through Bioconductor would make it easier to find, and give it extra credibility in the flow cytometry community. However, it would need go through another set of reviews and meet all of Bioconductor's requirements. I ran `R CMD BiocCheck .` on the package directory, and it flagged a few additional issues (see the [Bioconductor package submission guidelines](http://bioconductor.org/developers/package-submission/)).

**Publication strategy**: In the comments above, you mentioned submitting the package to JOSS. Of course the publication strategy is completely up to you; but I wanted to mention that I don't think this would be the best place for this package. JOSS papers are very short papers describing the purpose of the package only -- which is fine for many packages -- but for this package, I think it would be more useful to have a somewhat longer paper describing some of the key functionality, maybe along with a short workflow, and examples of visualizations. This would help convince flow cytometry users who are currently using commercial platforms (Cytobank, FlowJo, etc) that CytoRSuite could be a good alternative for them. Some possible venues for this type of paper could be the [Bioconductor gateway in F1000Research](https://f1000research.com/gateways/bioconductor), or more traditional flow cytometry journals like Cytometry Part A.
