### 2019-04-16 #281 @DillonHammill
Thanks for your review @lmweber!

Sorry the `testing` branch was not working as expected - I am working on some major updates and have not yet had time to push all the changes. I am working hard to refactor a lot of the code so I can implement `mockery` to improve the test coverage for the package.

I will need a week or so to finish implementing all of these new features and changes. I will merge them back to the `master` and let you know when everything is ready for a second look.
