### 2019-04-30 #281 @jacobpwagner
@geanders, it looks fixed in the `testing` branch, so as soon as that gets merged to `master` it should be good. I can hold off and look over everything once more after @DillonHammill has merged this batch of changes to `master`.
