### 2019-05-03 #281 @jacobpwagner
@DillonHammill I think you've seen that the API for `flowCore`, `flowWorkspace`, and `openCyto` is getting a bit of an overhaul, but it's mostly just method renaming with the old names being deprecated. The changes were merged to Bioconductor after the 3.9 branch was created, so for now it shouldn't affect users installing those dependencies following your installation instructions. However, if users install the dependencies from GitHub, they will get deprecation warnings from some of the methods you call. Anyway, I'd be happy to put together a PR just changing over the relevant method calls in your code that you could merge sometime before the next Bioconductor release.
