### 2019-05-03 #281 @DillonHammill
Thanks @jacobpwagner. Sorry for the delay, I am doing a complete overhaul of CytoRSuite to add some coding conventions to make it easier to add new features and find erroneous code. This is taking longer than expected but I will do my best to have it completed on the coming week. Looking forward to sharing the improvements soon. I will let you know when everything is merged to `master` and ready for a second look.
