### 2019-05-08 #281 @jacobpwagner
@DillonHammill , see https://github.com/DillonHammill/CytoRSuite/pull/17, which is based off of your `master` branch. I can help resolve conflicts if you merge in your `testing` branch.
