### 2019-12-09 #281 @DillonHammill
@geanders : Unfortunately, due to recent changes in the underlying data structures in the RGLab dependencies of this package (https://github.com/RGLab/flowWorkspace/issues/294 ), I am going to need some more time to get everything on par with the RGLab suite of packages. Everything is ready on my end, I will just need to make some changes as to how the data is handled internally to ensure that the package behaves as it used to. I don't see a point in re-submitting the package until it is up to date with the latest RGLab packages. I will do my best to make all the necessary changes over the next week or so. Apologies for further delaying this re-submission, but I think it is essential that I make these changes now. I am happy for you to put this hold whilst I get everything ready.
