### 2021-02-18 #281 @noamross
Hello @DillonHammill, I am Noam and I'm going to be taking over as handling editor for **CytoRSuite** from @geanders. I'll take this opportunity to check on your status. I see that this review is in "holding" as you were working on some considerable changes last year before starting review.  Do you have any updates?
